core: wifi: Move type classes to `typedefs.py`
